categories/web/web-easy-csp_xss/give/requirements.txt
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
werkzeug
pyjwt
redis